+++ deepl-java/src/main/java/com/deepl/api/Translator.java
@@ -85,7 +85,7 @@ public class Translator {
    */
   private String constructUserAgentString(boolean sendPlatformInfo, AppInfo appInfo) {
     StringBuilder sb = new StringBuilder();
-    sb.append("deepl-java/1.1.0");
+    sb.append("deepl-java/1.2.0");
     if (sendPlatformInfo) {
       sb.append(" (");
       Properties props = System.getProperties();