+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+* Script to check our source code for license headers and a step for them in the CI.
+
 ## [1.1.0] - 2023-01-26
 ### Added
 * Add example maven project using this library.
@@ -71,6 +75,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Initial version.
 
 
+[Unreleased]: https://github.com/DeepLcom/deepl-java/compare/v1.1.0...HEAD
 [1.1.0]: https://github.com/DeepLcom/deepl-java/compare/v1.0.1...v1.1.0
 [1.0.1]: https://github.com/DeepLcom/deepl-java/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/DeepLcom/deepl-java/compare/v0.2.1...v1.0.0